refactor(transformer): move `BoundIdentifier` into helpers (#3610)
Pure refactor. Move `BoundIdentifier` into helpers module so it can be reused by other transforms.
